+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.1.7 / Unreleased
 
 * Command list ips added
+* Command actions added
 * OSX install: `brew cask install onlabs`
 * bash tab completion added
 * Volume and Snapshot sizes are displayed in GB